+++ development.ts
@@ -4,7 +4,6 @@
 
 export type {ContextOptions} from './testing/execution';
 export type {ExecuteOptions} from './testing/execution';
-export type {ExecuteSyncOptions} from './testing/execution';
 export {executeFormulaFromPackDef} from './testing/execution';
 export {executeMetadataFormula} from './testing/execution';
 export {executeSyncFormulaFromPackDef} from './testing/execution';
